SwiftPM impliment missing PVLibrary swift stubs

+++ PVLibrary/Sources/PVLibrary/RealmPlatform/RomDatabase.swift
@@ -112,7 +112,7 @@ public final class RealmConfiguration {
                         return
                     }
 
-                    if let md5 = FileManager.default.md5ForFile(atPath: fullPath.path, fromOffset: offset), !md5.isEmpty {
+                    if let md5 = FileManager.default.md5ForFile(atPath: fullPath.path, fromOffset: Int64(offset)), !md5.isEmpty {
                         newObject!["md5Hash"] = md5
                         counter += 1
                     } else {
